CameraTeam/requirements.txt
matplotlib==3.10.0
numpy==2.2.1
pybullet==3.2.6
PyQt6==6.8.0
PyQt6_sip==13.8.0
PyYAML==6.0.2
PyYAML==6.0.2
Quartz==0.0.1.dev0
